Addresses Issue #180: Schema changes are not correctly re-synced
It's still not clear to me what's going wrong, but it may be related to
timezones because the version was being stored as MySQL date. This
commit adds a version field which uses a long rather than date type.

So we're still vulnerable if the server's clock is wrong, but at least
MySQL does not get involved with date conversion.