Fixes Issue #182: disables LocationType combo for already-saved Activities
Changing the locationType of an Activity after it's been created and
potentially had sites linked has not been implemented, and may be to
complex to consider. (Especially when considering what to do with sites
that have been entered offline and then sent to the server after the
locationType has changed -- eek)

To make this clearer for users, the combo is disabled after the activity
has been saved to make it clear(er) that you have to create a new activity.