[Fixes issue #38001821] Browser doesn't recognize exported format
The export process has been improved to incorporate a human-friendly filename
and time stamp into the exported filename.

Also, to work around the problems with the Content-disposition encoding
(see http://stackoverflow.com/questions/93551/how-to-encode-the-filename-parameter-of-content-disposition-header-in-http)
the filename is incorporated into the url, so the browser is pointed to
www.activityinfo.org/generated/XYZ1324234AFFADFD/My+PowerPoint+Pres.ppt

Also, took this opportunity to clarify the class naming:
xxImageStorageProvider -> xxStorageProvider

as these classes are now used for more than just images.