[Fixes Issue #38013577] Fixes renderign of Google Base maps in Itext renderers
Also, RTF renderer now renders maps as PNG images and embeds the PNG in to the
RTF document rather than trying to embed the map tiles into EMF images. Support
for embedded images within EMF seems pretty sketchy, the PNG are a safer option
for now. It is still disappointing that symbols can't be rendered as vectors within
RTF exports, but for this I think we will need to write a proper .docx exporter,
which would probably be a major project.