Fixes appengine-web.xml descriptor
We use maven filtering to set the version of the application
at build time. When running in eclipse, however, the Google Eclipse
Plugin doesn't like ${gae.application.version} as a version id.

To fix this, we have one appengine-web.xml for development in war/WEB-INF
which uses a fixed version number of "dev", and a second for
deployment in src/main/webapp/WEB-INF that uses maven filtering
to set the right version at runtime.