[AI-214] Store, version admin geometry in database
Admin entity geometry is now stored in the MySQL database and
can be fully queried from within AI and via the API as either
GeoJSON, Google Polylines, or as tile sets.

Currently the tile rendering is very primitive, need to see if
we can get better font rendering on AppEngine or whether we need
to do the rendering offline and upload the pre-rendered tiles to
the server.